Functional Serialization
Serializer has been updated. It is now two nodes: a UI node and a Serializer node to interface with the other nodes. Serialization framework is here but has only been implemented for the LowerTimeline. Deserialization currently not allowed from any other paths besides saved path. This is a simple fix though.